requirements.txt
customtkinter
face_recognition
opencv-python
dlib
pyttsx3
pandas
pillow
